Rework NodeAppConfig.peers to return Vector[Peer] rather than Vector[String]
