+++ node/src/main/scala/org/bitcoins/node/util/BitcoinSNodeUtil.scala
@@ -1,5 +1,9 @@
 package org.bitcoins.node.util
 
+import org.bitcoins.core.api.node.Peer
+import org.bitcoins.core.util.NetworkUtil
+import org.bitcoins.node.config.NodeAppConfig
+
 import scala.util.Random
 
 object BitcoinSNodeUtil {
@@ -18,4 +22,19 @@ object BitcoinSNodeUtil {
     */
   def createActorName(className: Class[_]): String =
     createActorName(className.getSimpleName)
+
+  def stringsToPeers(addresses: Vector[String])(implicit
+      nodeAppConfig: NodeAppConfig): Vector[Peer] = {
+    val formatStrings = addresses.map { s =>
+      //assumes strings are valid, todo: add util functions to check fully for different addresses
+      if (s.count(_ == ':') > 1 && s(0) != '[') //ipv6
+        "[" + s + "]"
+      else s
+    }
+    val inetSockets = formatStrings.map(
+      NetworkUtil.parseInetSocketAddress(_, nodeAppConfig.network.port))
+    val peers =
+      inetSockets.map(Peer.fromSocket(_, nodeAppConfig.socks5ProxyParams))
+    peers
+  }
 }
